pyproject.toml
[project]
name = "acs_cli_mi"
version = "0.1.3"
description = "CLI for querying Census ACS 5-year data for Michigan counties"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "typer>=0.24.1",
    "rich>=14.0.0",
]

[project.scripts]
acs-cli-mi = "acs_cli.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/acs_cli"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "respx>=0.22.0",
]
